Route: PUT /person/<id>/pay
If the *person* with the given ``id`` already have an active legacy profile,
it's lock_date will be extended (from this date) by number of
LEGACY_EXTEND_DAYS (from app.config).

If the person is not assigned a legacy profile, new one will be created and
lock_date will be set to LEGACY_LOCK_DAYS from today.

In either case the status of the legacy profile will be set to ACTIVE.

Only persons in UNPAID or ACTIVE status can make payments.